refactor(ui): collapse sidebar inheritance tower
