+++ ui/src/components/app-sidebar.ts
@@ -1,22 +1,8 @@
-import { html, nothing } from "lit";
+import { html, nothing, type PropertyValues, type TemplateResult } from "lit";
 import { state } from "lit/decorators.js";
-import type { GatewayControlUiPluginTab } from "../api/gateway.ts";
-import {
-  serializeSidebarEntry,
-  type NavigationRouteId,
-  type SidebarZoneEntry,
-} from "../app-navigation.ts";
-import { pathForRoute } from "../app-route-paths.ts";
-import { sessionHasPendingApproval } from "../app/approval-presentation.ts";
+import type { SessionObserverDigest } from "../../../packages/gateway-protocol/src/schema/sessions.js";
 import { beginNativeWindowDragFromTopInset } from "../app/native-window-drag.ts";
-import { controlUiPublicAssetPath } from "../app/public-assets.ts";
-import { readPresenceEntries, resolveCurrentSelfUser } from "../app/user-profile.ts";
-import { CONTROL_UI_BUILD_INFO } from "../build-info.ts";
-import { t } from "../i18n/index.ts";
-import { normalizeAgentLabel, resolveAgentTextAvatar } from "../lib/agents/display.ts";
-import { resolveAgentAvatarUrl } from "../lib/avatar.ts";
 import { BoardAvailabilityController } from "../lib/board/availability-controller.ts";
-import { sessionHasBoard } from "../lib/board/provider.ts";
 import "./menu-surface.ts";
 import "./session-menu.ts";
 import "./sidebar-agent-card.ts";
@@ -27,17 +13,35 @@ import "./theme-mode-toggle.ts";
 import "./tooltip.ts";
 import { isGatewayMethodAdvertised } from "../lib/gateway-methods.ts";
 import { createIdleImport } from "../lib/idle-import.ts";
-import { searchForSession } from "../lib/sessions/index.ts";
-import { areUiSessionKeysEquivalent, normalizeAgentId } from "../lib/sessions/session-key.ts";
-import { pluginTabKey } from "../pages/plugin/route.ts";
+import { normalizeAgentId } from "../lib/sessions/session-key.ts";
+import { SubscriptionsController } from "../lit/subscriptions-controller.ts";
+import { sidebarPluginTabs } from "./app-sidebar-nav-menus.ts";
 import {
-  renderSidebarPluginTab,
-  shouldHandleNavigationClick,
-  sidebarPluginTabs,
-} from "./app-sidebar-nav-menus.ts";
-import { AppSidebarSessionListElement } from "./app-sidebar-session-list.ts";
-import type { SidebarRecentSession } from "./app-sidebar-session-types.ts";
-import type { SidebarWorkboardBoard } from "./app-sidebar-workboard.ts";
+  renderAppSidebarAttention,
+  renderAppSidebarBrand,
+  renderAppSidebarFooterBar,
+  renderAppSidebarHomeRow,
+  renderAppSidebarPagesHead,
+  renderAppSidebarPluginTabEntry,
+  renderAppSidebarZoneEntry,
+} from "./app-sidebar-render.ts";
+import type { CatalogSessionMenuRequest } from "./app-sidebar-session-catalogs.ts";
+import { renderSessionList } from "./app-sidebar-session-list-render.ts";
+import type {
+  SidebarNarrationSyncInput,
+  SidebarSessionNarrationController,
+} from "./app-sidebar-session-narration.ts";
+import { AppSidebarSessionNavigationElement } from "./app-sidebar-session-navigation.ts";
+import {
+  renderSessionTree,
+  type SessionListHost,
+  visibleSessionChildren,
+} from "./app-sidebar-session-row-render.ts";
+import {
+  loadStoredSidebarCatalogGrouping,
+  storeSidebarCatalogGrouping,
+  type SidebarRecentSession,
+} from "./app-sidebar-session-types.ts";
 import { icons } from "./icons.ts";
 import {
   LOBSTER_LOGO_VISIT_EVENT,
@@ -46,12 +50,9 @@ import {
   resolveLobsterRunOutcome,
   type LobsterLogoVisitDetail,
 } from "./lobster-pet-contract.ts";
-import { redactLoginFailureError } from "./login-gate.ts";
-import { renderOfflineSidebarStatus, renderSessionRowBadges } from "./session-row-badges.ts";
-
-const PALETTE_SHORTCUT = /Mac|iP(hone|ad|od)/i.test(globalThis.navigator?.platform ?? "")
-  ? "⌘K"
-  : "Ctrl K";
+import { SessionOrganizerController } from "./session-organizer-controller.ts";
+import { renderSessionCreatorFilter } from "./session-owner-chip.ts";
+import { SidebarMenusController } from "./sidebar-menus-controller.ts";
 // The shared loader retries transient chunk failures online; a deploy-pruned
 // chunk still stays off until reload when that retry fails, by design.
 const sidebarChromeImport = createIdleImport(() =>
@@ -61,11 +62,27 @@ const sidebarChromeImport = createIdleImport(() =>
   ]),
 );
 
-class AppSidebar extends AppSidebarSessionListElement {
+class AppSidebar extends AppSidebarSessionNavigationElement implements SessionListHost {
+  @state() sidebarNarrationLines: ReadonlyMap<string, string> = new Map();
+  @state() sidebarObserverDigests: ReadonlyMap<string, SessionObserverDigest> = new Map();
+
+  override readonly sessionOrganizer = new SessionOrganizerController(this);
+  override readonly sidebarMenus = new SidebarMenusController(this);
+
+  // Lazy: the controller pulls core token-suppression modules that must stay
+  // out of the startup chunk (QA smoke startup-JS budget). It loads on the
+  // first update with the preference enabled; earlier events are safely
+  // dropped because the controller aligns from cumulative snapshots.
+  private narration: SidebarSessionNarrationController | null = null;
+  private narrationLoad: Promise<void> | null = null;
+  private readonly narrationSubscriptions = this.createNarrationSubscriptions();
+
+  @state() protected catalogProjectGrouping = loadStoredSidebarCatalogGrouping();
   @state() private logoVisit: LobsterLogoVisitDetail | null = null;
 
   constructor() {
     super();
+    void this.narrationSubscriptions;
     void new BoardAvailabilityController(
       this,
       () => {
@@ -101,6 +118,95 @@ class AppSidebar extends AppSidebarSessionListElement {
     this.addEventListener(LOBSTER_LOGO_VISIT_EVENT, this.handleLogoVisit as EventListener);
   }
 
+  private createNarrationSubscriptions(): SubscriptionsController {
+    const subscriptions = new SubscriptionsController(this);
+    subscriptions.effect(
+      () => this.context?.gateway,
+      (gateway) => gateway.subscribeEvents((event) => this.narration?.handleEvent(event)),
+    );
+    return subscriptions;
+  }
+
+  override disconnectedCallback() {
+    this.narration?.disconnect();
+    super.disconnectedCallback();
+  }
+
+  protected override willUpdate(changed: PropertyValues<this>) {
+    super.willUpdate(changed);
+    // A fresh draft must be visible where it will live: genuinely expand a
+    // collapsed Threads section (persisted) instead of overriding at render
+    // time, so the header toggle keeps matching the visible state.
+    if (
+      changed.has("draftSessionAgentId") &&
+      this.draftSessionAgentId &&
+      this.collapsedSessionSections.has("ungrouped")
+    ) {
+      this.sessionOrganizer.toggleSection("ungrouped");
+    }
+  }
+
+  override updated(changedProperties: PropertyValues<this>) {
+    super.updated(changedProperties);
+    if (!this.narration) {
+      if (this.sidebarLiveActivity) {
+        this.ensureNarrationController();
+      }
+      return;
+    }
+    this.narration.sync(this.narrationSyncInput());
+  }
+
+  private visibleNarrationRowsInOrder(): SidebarRecentSession[] {
+    const rows: SidebarRecentSession[] = [];
+    const append = (session: SidebarRecentSession) => {
+      rows.push(session);
+      if (this.isSessionChildrenExpanded(session)) {
+        visibleSessionChildren({
+          session,
+          fullyShownChildSessionKeys: this.fullyShownChildSessionKeys,
+        }).forEach(append);
+      }
+    };
+    this.visibleSessionRowsInOrder().forEach(append);
+    return rows;
+  }
+
+  private narrationSyncInput(): SidebarNarrationSyncInput {
+    const gateway = this.context?.gateway.snapshot;
+    return {
+      enabled: this.sidebarLiveActivity,
+      connected: this.connected && gateway?.connected === true,
+      connectionIdentity: gateway?.client ?? null,
+      source: this.context?.sessions ?? null,
+      rows: this.visibleNarrationRowsInOrder(),
+      openSessionKey: this.activeRouteId === "chat" ? this.getRouteSessionKey() : "",
+      agentId: this.selectedAgentIdForSessions(),
+    };
+  }
+
+  private ensureNarrationController(): void {
+    if (this.narration || this.narrationLoad) {
+      return;
+    }
+    this.narrationLoad = import("./app-sidebar-session-narration.ts").then((module) => {
+      this.narrationLoad = null;
+      // The element may have left the DOM while the chunk loaded.
+      if (!this.isConnected) {
+        return;
+      }
+      this.narration = new module.SidebarSessionNarrationController(
+        (lines) => {
+          this.sidebarNarrationLines = lines;
+        },
+        (digests) => {
+          this.sidebarObserverDigests = digests;
+        },
+      );
+      this.narration.sync(this.narrationSyncInput());
+    });
+  }
+
   override connectedCallback() {
     super.connectedCallback();
     // The decorative pet's large module stays out of startup and upgrades in place.
@@ -119,329 +225,144 @@ class AppSidebar extends AppSidebarSessionListElement {
     this.logoVisit = detail.phase === "out" ? null : detail;
   };
 
-  private renderBrand() {
-    const collapseLabel = t("nav.collapse");
-    const { activeId: cardAgentId, agent: cardAgent, agents: cardAgents } = this.activeChipAgent();
-    const menuUnread = cardAgents.some((entry) => {
-      const agentId = normalizeAgentId(entry.id);
-      return agentId !== cardAgentId && this.agentUnreadCount(agentId) > 0;
-    });
-    const cardName = cardAgent ? normalizeAgentLabel(cardAgent) : cardAgentId;
-    const approvalCount =
-      this.sessionData.approvalBadgeSnapshot().agentCounts.get(cardAgentId) ?? 0;
-    const cardAvatarText =
-      (cardAgent ? resolveAgentTextAvatar(cardAgent) : null) ??
-      (cardName || cardAgentId).slice(0, 1).toUpperCase();
-    // The sidebar action follows gateway availability; collapsed native chrome
-    // keeps its separate offline-tolerant ⌘N mirror.
-    return html`
-      <div class="sidebar-brand">
-        <openclaw-sidebar-agent-card
-          .agentName=${cardName}
-          .avatarUrl=${cardAgent ? resolveAgentAvatarUrl(cardAgent) : null}
-          .avatarText=${cardAvatarText}
-          .subtitle=${this.agentChipSubtitle(cardAgentId)}
-          .menuOpen=${this.sidebarMenus.agentMenuPosition !== null}
-          .menuUnread=${menuUnread}
-          .approvalCount=${approvalCount}
-          .switcherAvailable=${cardAgents.length > 1}
-          .onToggleMenu=${(trigger: HTMLElement) => this.sidebarMenus.toggleAgentMenu(trigger)}
-        ></openclaw-sidebar-agent-card>
-        <div class="sidebar-brand__actions">
-          <openclaw-tooltip
-            .content=${this.connected
-              ? t("chat.runControls.newSession")
-              : t("chat.runControls.newSessionDisconnected")}
-          >
-            <button
-              class="sidebar-brand__icon sidebar-brand__new-thread"
-              type="button"
-              @click=${() => this.onOpenNewSession?.(this.expandedAgentId())}
-              aria-label=${t("chat.runControls.newSession")}
-              ?disabled=${!this.connected}
-            >
-              ${icons.plus}
-            </button>
-          </openclaw-tooltip>
-          ${this.renderSearch()}
-          <openclaw-tooltip .content=${`${collapseLabel} (⌘B)`}>
-            <button
-              class="sidebar-brand__icon sidebar-brand__collapse"
-              type="button"
-              @click=${() => this.onToggleSidebar?.()}
-              aria-label=${collapseLabel}
-              aria-expanded="true"
-            >
-              ${icons.panelLeftClose}
-            </button>
-          </openclaw-tooltip>
-        </div>
-      </div>
-    `;
+  startSessionDrag(session: SidebarRecentSession): void {
+    this.sessionOrganizer.startSessionDrag(session);
   }
 
-  /** Home: the first page. Opens the rolling main session on its saved face. */
-  private renderHomeRow() {
-    const agentId = this.activeChipAgent().activeId;
-    const mainKey = this.selectedAgentMainSessionKey(agentId);
-    const mainRow = this.mainSessionRow(agentId);
-    const approvalNeeded = sessionHasPendingApproval(
-      this.sessionData.approvalBadgeSnapshot(),
-      mainKey,
-    );
-    const outboxCount = this.outboxCountForSessionKey(mainKey);
-    const active =
-      this.activeRouteId === "chat" &&
-      areUiSessionKeysEquivalent(this.getRouteSessionKey(), mainKey);
-    const stateBadge = mainRow?.hasActiveRun
-      ? html`<openclaw-tooltip .content=${t("sessionsView.activeRun")}>
-          <span
-            class="session-run-spinner"
-            role="img"
-            aria-label=${t("sessionsView.activeRun")}
-          ></span>
-        </openclaw-tooltip>`
-      : mainRow?.unread === true && !active
-        ? html`<span
-            class="session-unread-dot"
-            role="img"
-            aria-label=${t("sessionsView.unread")}
-          ></span>`
-        : nothing;
-    return html`
-      <a
-        href=${`${pathForRoute("chat", this.basePath)}${searchForSession(mainKey)}`}
-        class="nav-item nav-item--home ${active ? "nav-item--active" : ""}"
-        aria-current=${active ? "page" : nothing}
-        @click=${(event: MouseEvent) => {
-          if (!shouldHandleNavigationClick(event)) {
-            return;
-          }
-          event.preventDefault();
-          this.openMainSession(agentId);
-        }}
-      >
-        <span class="nav-item__icon" aria-hidden="true">${icons.home}</span>
-        <span class="nav-item__text">${t("nav.home")}</span>
-        ${sessionHasBoard(mainKey)
-          ? html`<openclaw-tooltip .content=${t("sessionsView.dashboardAvailable")}>
-              <span
-                class="sidebar-board-glyph"
-                role="img"
-                aria-label=${t("sessionsView.dashboardAvailable")}
-                >${icons.layoutDashboard}</span
-              >
-            </openclaw-tooltip>`
-          : nothing}
-        ${stateBadge !== nothing || approvalNeeded || outboxCount > 0
-          ? html`<span class="nav-item__state sidebar-home-session-states">
-              ${stateBadge}
-              ${approvalNeeded
-                ? html`<openclaw-tooltip .content=${t("sessionsView.approvalNeeded")}>
-                    <span
-                      class="session-approval-badge"
-                      role="img"
-                      aria-label=${t("sessionsView.approvalNeeded")}
-                      >${icons.alertTriangle}</span
-                    >
-                  </openclaw-tooltip>`
-                : nothing}
-              ${renderSessionRowBadges({ hasAutomation: false, outboxCount })}
-            </span>`
-          : nothing}
-      </a>
-    `;
+  finishSessionDrag(): void {
+    this.sessionOrganizer.finishSessionDrag();
   }
 
-  /** "Pages" header: the customize affordance opens the pages menu (all
-      routes navigable, pin editor behind it) that used to hide behind More. */
-  private renderPagesHead() {
-    return html`
-      <div class="sidebar-nav__head">
-        <span class="sidebar-recent-sessions__label-text">${t("nav.pages")}</span>
-        <button
-          type="button"
-          class="sidebar-nav__head-action"
-          aria-haspopup="menu"
-          aria-expanded=${String(this.sidebarMenus.moreMenuPosition !== null)}
-          aria-label=${t("nav.customize")}
-          @click=${(event: MouseEvent) =>
-            this.sidebarMenus.toggleMoreMenu(event.currentTarget as HTMLElement)}
-        >
-          ${icons.penLine}
-        </button>
-      </div>
-    `;
+  toggleSessionPin(session: SidebarRecentSession): void {
+    void this.sessionOrganizer.patchSession(session, { pinned: !session.pinned });
   }
 
-  /** Zone 5: product chrome recedes to one slim footer bar. */
-  private renderFooterBar() {
-    const reconnecting = t("connection.reconnecting");
-    const selfUser = this.connected
-      ? resolveCurrentSelfUser({
-          snapshotUser: this.context?.gateway.snapshot.selfUser,
-          presenceEntries: readPresenceEntries(this.sessionData.presencePayload),
-          presenceInstanceId: this.sessionData.presenceInstanceId,
-        })
-      : null;
-    const selfLabel = selfUser?.name ?? selfUser?.email ?? selfUser?.id;
-    return html`
-      <div class="sidebar-footer-bar">
-        <span class="sidebar-brand__logo-slot sidebar-footer-bar__logo">
-          <img
-            class="sidebar-brand__logo ${this.logoVisit ? "sidebar-brand__logo--vacated" : ""}"
-            src=${controlUiPublicAssetPath("apple-touch-icon.png", this.basePath)}
-            alt=""
-            aria-hidden="true"
-          />
-          <openclaw-lobster-logo-standin .visit=${this.logoVisit}></openclaw-lobster-logo-standin>
-        </span>
-        ${selfUser && selfLabel
-          ? html`<openclaw-tooltip .content=${selfLabel}>
-              <button
-                type="button"
-                class="sidebar-footer-bar__identity"
-                aria-label=${t("profilePage.identity.openSettings", { name: selfLabel })}
-                @click=${() => this.onNavigate?.("profile", { hash: "#settings-profile-identity" })}
-              >
-                <openclaw-viewer-avatar
-                  .user=${{ ...selfUser, watchedSessions: [] }}
-                  variant="footer"
-                ></openclaw-viewer-avatar>
-                <span class="sidebar-footer-bar__identity-name">${selfLabel}</span>
-              </button>
-            </openclaw-tooltip>`
-          : nothing}
-        <openclaw-viewer-facepile
-          .presencePayload=${this.sessionData.presencePayload}
-          .selfInstanceId=${this.sessionData.presenceInstanceId}
-          .buildInfo=${CONTROL_UI_BUILD_INFO}
-          .gatewayVersion=${this.gatewayVersion}
-          .maxVisible=${5}
-          variant="footer"
-        ></openclaw-viewer-facepile>
-        <openclaw-sidebar-build-chip
-          .basePath=${this.basePath}
-          .gatewayVersion=${this.gatewayVersion}
-          .onNavigate=${(routeId: "about") => this.onNavigate?.(routeId)}
-        ></openclaw-sidebar-build-chip>
-        ${this.offline
-          ? renderOfflineSidebarStatus({
-              queuedOutboxCount: this.queuedOutboxCount,
-              reconnecting,
-              title: this.lastError ? redactLoginFailureError(this.lastError) : reconnecting,
-              onRetry: () => this.onRetryConnect?.(),
-            })
-          : nothing}
-        <openclaw-tooltip .content=${t("nav.settings")}>
-          <button
-            type="button"
-            class="sidebar-footer-bar__settings"
-            aria-label=${t("nav.settings")}
-            @click=${() => this.onNavigate?.("config")}
-          >
-            ${icons.settings}
-          </button>
-        </openclaw-tooltip>
-      </div>
-    `;
-  }
-
-  private renderSearch() {
-    const tooltip = `${t("chat.openCommandPalette")} (${PALETTE_SHORTCUT})`;
-    return html`
-      <openclaw-tooltip .content=${tooltip}>
-        <button
-          type="button"
-          class="sidebar-brand__icon sidebar-search"
-          ?disabled=${!this.onOpenPalette}
-          aria-label=${t("chat.openCommandPalette")}
-          @click=${() => this.onOpenPalette?.()}
-        >
-          ${icons.search}
-        </button>
-      </openclaw-tooltip>
-    `;
-  }
-
-  private renderSidebarZoneEntry(
-    entry: SidebarZoneEntry,
-    sessionRows: ReadonlyMap<string, SidebarRecentSession>,
-    workboardRows: ReadonlyMap<string, SidebarWorkboardBoard>,
-  ) {
-    if (
-      (entry.type === "route" && !this.sidebarMenus.isRouteEnabled(entry.route)) ||
-      (entry.type === "workboard" && !this.sidebarMenus.isRouteEnabled("workboard"))
-    ) {
-      return nothing;
+  toggleSessionMenu(
+    session: SidebarRecentSession,
+    menuSession: SidebarRecentSession,
+    trigger: HTMLElement,
+  ): void {
+    if (this.sidebarMenus.sessionMenu?.session.key === session.key) {
+      this.sidebarMenus.closeSessionMenu();
+      return;
     }
-    const serialized = serializeSidebarEntry(entry);
-    const dropPosition =
-      this.sessionOrganizer.sidebarZoneDropTarget?.entry === serialized
-        ? this.sessionOrganizer.sidebarZoneDropTarget.position
-        : null;
-    const content =
-      entry.type === "route"
-        ? this.sidebarMenus.renderRoute(entry.route)
-        : entry.type === "workboard"
-          ? workboardRows.has(entry.boardId)
-            ? this.renderWorkboardBoard(workboardRows.get(entry.boardId)!)
-            : nothing
-          : sessionRows.has(entry.key)
-            ? this.renderPinnedSidebarSession(sessionRows.get(entry.key)!)
-            : nothing;
-    const draggable = entry.type === "route" || entry.type === "workboard";
-    return html`
-      <div
-        class="sidebar-zone-entry ${dropPosition
-          ? `sidebar-zone-entry--drop-${dropPosition}`
-          : ""} ${this.sessionOrganizer.draggingSidebarEntry === serialized
-          ? "sidebar-zone-entry--dragging"
-          : ""}"
-        data-sidebar-entry=${serialized}
-        draggable=${draggable ? "true" : "false"}
-        @dragstart=${entry.type === "route"
-          ? (event: DragEvent) => this.sessionOrganizer.startSidebarRouteDrag(event, entry.route)
-          : entry.type === "workboard"
-            ? (event: DragEvent) =>
-                this.sessionOrganizer.startSidebarWorkboardDrag(event, entry.boardId)
-            : nothing}
-        @dragend=${draggable ? () => this.sessionOrganizer.finishSidebarEntryDrag() : nothing}
-        @dragover=${(event: DragEvent) =>
-          this.sessionOrganizer.handleSidebarZoneDragOver(event, serialized)}
-        @drop=${(event: DragEvent) =>
-          this.sessionOrganizer.handleSidebarZoneDrop(event, serialized)}
-      >
-        ${content}
-      </div>
-    `;
+    const rect = trigger.getBoundingClientRect();
+    this.sidebarMenus.openSessionMenu(menuSession, rect.right, rect.bottom + 4, trigger);
   }
 
-  private renderPluginTabEntry(tab: GatewayControlUiPluginTab) {
-    const ref = { pluginId: tab.pluginId, id: tab.id };
-    const key = pluginTabKey(ref);
-    return html`
-      <div class="sidebar-zone-entry" data-sidebar-entry=${`plugin:${key}`}>
-        ${renderSidebarPluginTab({
-          tab,
-          basePath: this.basePath,
-          active: this.activeRouteId === "plugin" && this.activePluginTabId === key,
-          onNavigate: (search) => this.onNavigate?.("plugin", { search }),
-        })}
-      </div>
-    `;
+  startSessionGroupDrag(group: string): void {
+    this.sessionOrganizer.startSessionGroupDrag(group);
   }
 
-  private renderWorkboardBoard(board: SidebarWorkboardBoard) {
-    const active = this.activeRouteId === "workboard" && this.activeWorkboardBoardId === board.id;
-    return (
-      this.workboardRenderers?.renderEntry({
-        board,
+  finishSessionGroupDrag(): void {
+    this.sessionOrganizer.finishSessionGroupDrag();
+  }
+
+  sectionDragOver(event: DragEvent, sectionId: string, group?: string): void {
+    this.sessionOrganizer.sectionDragOver(event, sectionId, group);
+  }
+
+  sectionDragLeave(event: DragEvent, sectionId: string, group?: string): void {
+    this.sessionOrganizer.sectionDragLeave(event, sectionId, group);
+  }
+
+  sectionDrop(event: DragEvent, sectionId: string, group?: string): void {
+    this.sessionOrganizer.sectionDrop(event, sectionId, group);
+  }
+
+  toggleSection(sectionId: string): void {
+    this.sessionOrganizer.toggleSection(sectionId);
+  }
+
+  handleSessionListDragOver(event: DragEvent): void {
+    this.sessionOrganizer.handleSessionListDragOver(event);
+  }
+
+  handleSessionListDragLeave(event: DragEvent): void {
+    this.sessionOrganizer.handleSessionListDragLeave(event);
+  }
+
+  handleSessionListDrop(event: DragEvent): void {
+    this.sessionOrganizer.handleSessionListDrop(event);
+  }
+
+  openNewSession(): void {
+    this.onOpenNewSession?.(this.expandedAgentId());
+  }
+
+  setVisibleSessionLimit(limit: number): void {
+    this.sessionData.setVisibleSessionLimit(limit);
+  }
+
+  dismissSessionMutationError(): void {
+    this.sessionData.dismissSessionMutationError();
+  }
+
+  toggleCatalogProjectGrouping(): void {
+    const next = this.catalogProjectGrouping === "project" ? "none" : "project";
+    storeSidebarCatalogGrouping(next);
+    this.catalogProjectGrouping = next;
+  }
+
+  openCatalogMenu(
+    request: CatalogSessionMenuRequest,
+    x: number,
+    y: number,
+    trigger?: HTMLElement,
+  ): void {
+    this.sidebarMenus.catalogMenu.open(request, x, y, trigger);
+  }
+
+  renderPinnedSidebarSession(session: SidebarRecentSession): TemplateResult {
+    return renderSessionTree({ host: this, session });
+  }
+
+  private renderSessions() {
+    const navigationState = this.getSessionNavigationState();
+    const visibleSessions = this.selectedAgentSessionRows(navigationState);
+    const expandedAgentId = this.expandedAgentId();
+    const liveRows = [
+      ...(this.sessionData.sessionsResult?.sessions ?? []),
+      ...Object.values(this.sessionData.sessionRowsByAgent).flat(),
+    ];
+    const sidebarRowsByKey = new Map<string, SidebarRecentSession>();
+    for (const row of liveRows) {
+      if (!sidebarRowsByKey.has(row.key)) {
+        sidebarRowsByKey.set(row.key, navigationState.toSidebarSession(row));
+      }
+    }
+    const { sections, expandedRows, visibleRows } = this.zonedVisibleSections(visibleSessions);
+    return renderSessionList({
+      host: this,
+      empty: visibleSessions.length === 0,
+      sections,
+      expandedRows,
+      visibleRowCount: visibleRows.length,
+      showDraft:
+        Boolean(this.draftSessionAgentId) &&
+        normalizeAgentId(this.draftSessionAgentId) === expandedAgentId,
+      creatorFilter: renderSessionCreatorFilter({
+        creators: this.sessionOwnershipVisible ? this.sessionCreatorOptions : [],
+        selectedId: this.sessionCreatorFilterActive ? this.sessionCreatorFilterId : null,
+        onChange: (creatorId) => {
+          this.sessionCreatorFilterId = creatorId;
+          void this.context?.sessions.setCreatorFilter(creatorId);
+        },
+      }),
+      catalogs: {
+        catalogs: this.sessionData.sessionCatalogs,
         basePath: this.basePath,
-        active,
-        onNavigate: (pathname) => this.onNavigate?.("workboard", { pathname }),
-      }) ?? nothing
-    );
+        routeSessionKey: this.activeRouteId === "chat" ? this.getRouteSessionKey() : "",
+        newSessionAgentId: expandedAgentId,
+        loadingMoreCatalogIds: this.sessionData.loadingMoreSessionCatalogIds,
+        projectGrouping: this.catalogProjectGrouping,
+        liveRows,
+        sidebarRowsByKey,
+        creatorId: this.activeSessionCreatorId,
+        catalogOpenTarget: this.catalogOpenTarget,
+        terminalAvailable: this.terminalAvailable,
+      },
+    });
   }
 
   override render() {
@@ -449,7 +370,7 @@ class AppSidebar extends AppSidebarSessionListElement {
     return html`
       <aside class="sidebar">
         <div class="sidebar-shell" @mousedown=${beginNativeWindowDragFromTopInset}>
-          ${this.renderBrand()}
+          ${renderAppSidebarBrand(this)}
           <div
             class="sidebar-shell__body sidebar-shell__body--scroll-${this.sessionData
               .sessionsScrollState}"
@@ -457,7 +378,7 @@ class AppSidebar extends AppSidebarSessionListElement {
               this.sessionData.updateSessionsScrollState(event.currentTarget as HTMLElement)}
           >
             <nav class="sidebar-nav" @contextmenu=${this.sidebarMenus.openCustomizeMenuFromContext}>
-              ${this.renderPagesHead()}
+              ${renderAppSidebarPagesHead(this)}
               <div
                 class="nav-section__items"
                 @dragover=${(event: DragEvent) =>
@@ -466,26 +387,24 @@ class AppSidebar extends AppSidebarSessionListElement {
                   this.sessionOrganizer.handleSidebarZoneDragLeave(event)}
                 @drop=${(event: DragEvent) => this.sessionOrganizer.handleSidebarZoneDrop(event)}
               >
-                ${this.renderHomeRow()}
+                ${renderAppSidebarHomeRow(this)}
                 ${sidebarZone.entries.map((entry) =>
-                  this.renderSidebarZoneEntry(
+                  renderAppSidebarZoneEntry(
+                    this,
                     entry,
                     sidebarZone.sessionRows,
                     sidebarZone.workboardRows,
                   ),
                 )}
                 ${sidebarPluginTabs(this.context?.gateway.snapshot.hello?.controlUiTabs).map(
-                  (tab) => this.renderPluginTabEntry(tab),
+                  (tab) => renderAppSidebarPluginTabEntry(this, tab),
                 )}
               </div>
             </nav>
             ${this.renderSessions()}
           </div>
           <div class="sidebar-shell__footer">
-            <openclaw-sidebar-attention
-              .onNavigate=${(routeId: NavigationRouteId) => this.onNavigate?.(routeId)}
-              .onOpenApprovals=${() => this.onOpenApprovals?.()}
-            ></openclaw-sidebar-attention>
+            ${renderAppSidebarAttention(this)}
             <openclaw-sidebar-update-card
               .updateAvailable=${this.updateAvailable}
               .updateRunning=${this.updateRunning}
@@ -512,7 +431,7 @@ class AppSidebar extends AppSidebarSessionListElement {
                   </div>
                 </openclaw-tooltip>`
               : nothing}
-            ${this.renderFooterBar()}
+            ${renderAppSidebarFooterBar(this, this.logoVisit)}
           </div>
         </div>
         ${this.sidebarMenus.renderCustomizeMenu()} ${this.sidebarMenus.renderMoreMenu()}
